fix: 버그 수정
- save_cropped_people 함수에서 이미지 저장 시 bounding box 로 이미지가 잘린 문제 해결
- full_frames 함수에서 이미지 저장 시 bounding box 로 이미지가 잘린 문제 해결
- edit_main.py 파일 삭제
- 코드 리팩토링

Closes #49
